+++ formatters/date.js
@@ -2,7 +2,7 @@ var moment = require('moment');
 
 
 /**
- * Format dates
+ * Format dates (DEPRECATED, use formatD instead)
  * 
  * @exampleContext {"lang":"en"}
  * @example ["20160131", "YYYYMMDD", "L"]
@@ -28,6 +28,44 @@ function convDate (d, patternIn, patternOut) {
   return d;
 }
 
+/**
+ * Format dates
+ * New since 1.2.0
+ * 
+ * @exampleContext {"lang":"en"}
+ * @example ["20160131", "L"]
+ * @example ["20160131", "LL"]
+ * @example ["20160131", "LLLL"]
+ * @example ["20160131", "dddd"]
+ * @example [1410715640, "LLLL"]
+ *
+ * @exampleContext {"lang":"fr"}
+ * @example ["2017-05-10T15:57:23.769561+03:00", "LLLL"]
+ * @example ["2017-05-10 15:57:23.769561+03:00", "LLLL"]
+ * @example ["20160131", "LLLL"]
+ * @example ["20160131", "dddd"]
+ *
+ * @exampleContext {"lang":"fr"}
+ * @example ["20160131", "dddd", "YYYYMMDD"]
+ * @example [1410715640, "LLLL", "X" ]
+ * 
+ * @param  {String|Number} d   date to format
+ * @param  {String} patternOut output format
+ * @param  {String} patternIn  [optional] input format, ISO8601 by default
+ * @return {String}            return formatted date
+ */
+function formatD (d, patternOut, patternIn) {
+  if (d !== null && typeof d !== 'undefined') {
+    moment.locale(this.lang);
+    if (patternIn) {
+      return moment(d + '', patternIn).format(patternOut);
+    }
+    return moment(d + '').format(patternOut);
+  }
+  return d;
+}
+
 module.exports = {
+  formatD : formatD,
   convDate : convDate
 };